clean: made getting a user info from firestore resuable
