+++ src/components/hooks/myCustomHooks.ts
@@ -1,6 +1,9 @@
+import { firestore } from 'firebase';
+import { IUser } from './../../@types/types';
 import { useHistory } from "react-router-dom";
 import { useSelector } from "react-redux";
-import React, { useEffect } from "react";
+import React, { useEffect, useState } from "react";
+import { createPortal } from 'react-dom';
 
 export const useToHomeIfAuthenticated = () => {
   const history = useHistory();
@@ -22,4 +25,23 @@ export const useToLandingIfNotAuthenticated = () => {
   useEffect(() => {
     isEmpty && history.push("/");
   }, isEmpty);
+}
+
+export const GetUserDataFromFirestoreByUid = (uid: string): IUser => {
+  const [user, setUser] = useState({ name: "", email: "" })
+
+  useEffect(() => {
+    const getuser = async () => {
+      const res = await firestore()
+        .collection("users")
+        .doc(uid)
+        .get();
+      const data = res.data();
+      const { name, email } = data;
+      setUser({ name, email });
+    };
+    getuser();
+  }, [uid]);
+
+  return user;
 }
