+++ src/components/pages/home/posts/Comment.tsx
@@ -1,14 +1,15 @@
 import React from "react";
 import styled, { css } from "styled-components";
+import { GetUserDataFromFirestoreByUid } from "../../../hooks/myCustomHooks";
 interface Props {
   uid: string;
   comment: string;
 }
 
-//todo make getCreater general
 const Comment = (props: Props) => {
   const { uid, comment } = props;
-  console.log("comment:", comment);
+  const user = GetUserDataFromFirestoreByUid(uid);
+  console.log("user:", user);
   return (
     <div>
       <h5>Comment</h5>
